requirements.txt
# requirements.txt atualizado
torch>=2.2.0
torchvision>=0.16.0
opencv-python>=4.7.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=9.5.0
python-dateutil>=2.8.0